Created Customer Service and Service Implementation
The Service implements two methods. The 'Record' method checks  the Customer model's attributes .
If a Record in the database has the same attributes as the Customer model, the method will throw an error message, because this means that the customer is already added to the database. Else the Customer model will be passed towards to the Customer DAO to be added to the database.The 'Read all' method returns all of the Customers from the database by the Customer DAO.